+++ tests/api/test_ossl_x509_pk.h
@@ -30,6 +30,7 @@ int test_wolfSSL_X509_PUBKEY_EC(void);
 int test_wolfSSL_X509_PUBKEY_DSA(void);
 int test_wolfSSL_X509_PUBKEY_get(void);
 int test_wolfSSL_X509_set_pubkey(void);
+int test_wolfSSL_X509_get0_pubkey(void);
 
 #define TEST_OSSL_X509_PK_DECLS                                                \
     TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_get_X509_PUBKEY),        \
@@ -37,6 +38,7 @@ int test_wolfSSL_X509_set_pubkey(void);
     TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_PUBKEY_EC),              \
     TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_PUBKEY_DSA),             \
     TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_PUBKEY_get),             \
-    TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_set_pubkey)
+    TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_set_pubkey),            \
+    TEST_DECL_GROUP("ossl_x509_pk", test_wolfSSL_X509_get0_pubkey)
 
 #endif /* WOLFCRYPT_TEST_OSSL_X509_PK_H */